% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_annotation}
\alias{add_annotation}
\title{add annotation such as the beat cycle starts to the onset data frame}
\usage{
add_annotation(df = NULL, annotation = NULL, time = NULL, reference = NULL)
}
\arguments{
\item{df}{data frame to be processed (required)}

\item{annotation}{Annotation (cycle starts) to be added}

\item{time}{Annotated times of the cycle starts to be added}

\item{reference}{The target column (Label.SD) for cycles in IEMP data}
}
\value{
dataframe with added column of cycles and cycle onsets
}
\description{
This function adds cycle start times from annotations to onsets
}
\seealso{
\code{\link{add_isobeats}}
}
