% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_isobeats.R
\name{add_isobeats}
\alias{add_isobeats}
\title{Add isochronous beats to the file}
\usage{
add_isobeats(
  df = NULL,
  instr = NULL,
  beat = NULL,
  beatlabel = "Isochronous.SD.Time"
)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{variable which contains the instruments}

\item{beat}{variable which contains the beats}

\item{beatlabel}{variable which labels the newly created beats}
}
\value{
dataframe with added column of isochronous beats and cycle numbers
}
\description{
Add isochronous beats to a onset datafame based on beat levels and average onset of the 1st beat
}
\seealso{
\code{\link{add_annotation}}
}
