% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asere_OU_2.R
\docType{data}
\name{Asere_OU_2}
\alias{Asere_OU_2}
\title{Asere_OU_2}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{Piece}{Name of the piece}
  \item{Label.SD}{Label of SD}
  \item{Beat.pos}{Beat position}
  \item{SD.pos}{SD}
  \item{Section}{Section of the piece}
  \item{Virtual.SD}{Virtual}
  \item{Tactus}{Tactus}
  \item{Tempo}{Tempo}
  \item{Clave}{Instrument name Clave}
  \item{Bass}{Instrument name Bass}
  \item{Guitar}{Instrument name Guitar}
  \item{Tres}{Instrument name Tres}
  \item{Bongo}{Instrument name Bongo }
  \item{Bell}{Instrument name Bell}
}
}
\usage{
Asere_OU_2
}
\description{
Asere example, refer to IEMP collection for full details.
}
\keyword{datasets}
