% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSS_IEMP.R
\docType{data}
\name{CSS_IEMP}
\alias{CSS_IEMP}
\title{CSS_IEMP}
\format{
A list of 5 containing individual pieces:
\describe{
  \item{names}{Name of the piece}
  \item{id}{ID}
  \item{genre}{Genre of the piece}
  \item{class}{corpus}
}
}
\usage{
CSS_IEMP
}
\description{
Cuban Salsa and Son mini corpus, refer to IEMP collection for full details.
}
\keyword{datasets}
