% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OSF_csv.R
\name{get_OSF_csv}
\alias{get_OSF_csv}
\title{Get csv file from OSF}
\usage{
get_OSF_csv(address = NULL, filename = NULL)
}
\arguments{
\item{address}{OSF address code () to be retrieved. Works for CSV files.}

\item{filename}{Filename (optional)}
}
\value{
data
}
\description{
This function gets CSV file from OSF using httr GET method.
}
