% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nPVI.R
\name{nPVI}
\alias{nPVI}
\title{Measure of durational variability of events}
\usage{
nPVI(df = NULL, instr = NULL)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{Instrument to be processed}
}
\value{
integer value
}
\description{
This function calculates of durational variability of events (nPVI).
}
\details{
This measure is borne out of language research. It has been noted that 
the variability of vowel duration is greater in stress- vs. syllable-timed 
languages (Grabe & Low, 2002). This measure accounts for the 
variability of durations and is also called "normalized Pairwise Variability 
Index" (nPVI). Patel & Daniele have applied it to music (2003) by comparing
whether the prosody of different languages is also reflected in music. There is 
a clear difference between a sample of works by French and English composers.
}
