% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsync-package.R
\docType{package}
\name{onsetsync-package}
\alias{onsetsync}
\alias{onsetsync-package}
\title{onsetsync: Analysis and Visualisation of Synchronisation of Music Onset Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helps to calculate and visualise onset data. The emphasis is on comparing onsets between performers in the same ensemble, in other words, how well the performers synchronise to each other. There are also some routines to estimate to periodicity of the onsets.
}
\author{
\strong{Maintainer}: Tuomas Eerola \email{tuomas.eerola@durham.ac.uk}

}
\keyword{internal}
