% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_add_isobeats.R
\name{onsetsynch_add_isobeats}
\alias{onsetsynch_add_isobeats}
\title{onsetsynch_add_isobeats}
\usage{
onsetsynch_add_isobeats(df, instr, beat)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{variable which contains the instruments}

\item{beat}{variable which contains the beats}
}
\value{
dataframe with added column of isochronous beats and cycle numbers
}
\description{
Add isochronous beats to a onset datafame based on beat levels and average onset of the 1st beat
}
