% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_by_dataset_plot.R
\name{onsetsynch_by_dataset_plot}
\alias{onsetsynch_by_dataset_plot}
\title{plot synchronies by beat structures}
\usage{
onsetsynch_by_dataset_plot(df, asynchronies, data, pcols = 1, box = FALSE,
  colour = "lightblue")
}
\arguments{
\item{df}{data frame to be processed}

\item{data}{Variable name where the beats are}

\item{pcols}{Number of columns for multiple plots (default 1)}

\item{box}{Do the graphics by boxplot}

\item{colour}{colour for the boxplot}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
