% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_by_pair_plot.R
\name{onsetsynch_by_pair_plot}
\alias{onsetsynch_by_pair_plot}
\title{plot synchronies by pairs}
\usage{
onsetsynch_by_pair_plot(df, bybeat = FALSE)
}
\arguments{
\item{df}{data frame to be processed}

\item{bybeat}{Optional (FALSE by default) which addes the beats to the plot}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instrument pairs.
}
