% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_by_pair_stats.R
\name{onsetsynch_by_pair_stats}
\alias{onsetsynch_by_pair_stats}
\title{stats for asynchronies by pairs}
\usage{
onsetsynch_by_pair_stats(df, bybeat = FALSE)
}
\arguments{
\item{df}{data frame to be processed}

\item{bybeat}{logical value whether beat information is used}
}
\value{
Statistics output (Mean, Standard deviation, T-stat, P value)
}
\description{
This function calculates simple t-test of the asynchronies of instruments (whether it differs from zero)
}
