% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_by_X_plot.R
\name{onsetsynch_by_X_plot}
\alias{onsetsynch_by_X_plot}
\title{plot synchronies by any other structure}
\usage{
onsetsynch_by_X_plot(df, meta = "empty", xlab = "Tempo")
}
\arguments{
\item{df}{data frame to be processed}

\item{meta}{Variable information (Pair of instrument names, for instance)}

\item{xlab}{label for X axis}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments across another variable
}
