% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_execute_pairs.R
\name{onsetsynch_execute_pairs}
\alias{onsetsynch_execute_pairs}
\title{Calculate asynchronies across all pairs of instruments.}
\usage{
onsetsynch_execute_pairs(df, instruments, n_samp = 100, bnum = 1, beats)
}
\arguments{
\item{df}{data frame to be processed}

\item{instruments}{Instrument names to be processed}

\item{n_samp}{Number of samples to be drawn from the pool of joint onsets}

\item{beats}{Beat structure to be included}

\item{b_num}{How many bootstraps are drawn}
}
\value{
Output contain asynchronies and beat levels
}
\description{
\code{onsetsynch_execute_pairs} is a meta-function that
executes calculations of asynchronies across all pairs of
instruments. One can speficy the same options as in
\code{onsetsynch_sample_paired}.
}
\seealso{
\code{\link{onsetsynch_sample_paired}}
}
