% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_to_BPM.R
\name{period_to_BPM}
\alias{period_to_BPM}
\title{Convert period in seconds to Beats Per Minute (BPM)}
\usage{
period_to_BPM(period = NULL, beats_per_minute = 60)
}
\arguments{
\item{period}{data frame to be processed (required)}

\item{beats_per_minute}{baseline assumption of how many beats there are in minute (default 60)}
}
\value{
Numeric output
}
\description{
Converts period in seconds to BPM
}
\examples{
bpm <- period_to_BPM(0.666)
print(bpm)
}
\seealso{
\code{\link{summarise_periodicity}},\code{\link{periodicity}}
}
