% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity_moments.R
\name{periodicity_moments}
\alias{periodicity_moments}
\title{Summarise onset periodicities by calculating the moments (mean, etc.)}
\usage{
periodicity_moments(data = NULL, method = "max")
}
\arguments{
\item{data}{data frame to be processed}

\item{method}{method of choosing the peaks (max default, or first)}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
\seealso{
\code{\link{periodicity}}
}
