% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_beat.R
\name{plot_by_beat}
\alias{plot_by_beat}
\title{Plot synchronies by beat structures}
\usage{
plot_by_beat(
  df = NULL,
  instr = NULL,
  beat = NULL,
  virtual = NULL,
  pcols = 1,
  griddeviations = FALSE,
  boxplot = FALSE,
  colour = "lightblue",
  colourpalette = "Set2",
  pointsize = 1
)
}
\arguments{
\item{df}{data frame to be processed, where the minimal requirements are:
(1) two or more instruments (`instr`), (2) beat sub-division (`beat`), and
(3) reference timing (`virtual`), which can be either isochronous OR mean times, 
possibly created by [add_isobeat()]. See vignette("minimal-representation", package = "onsetsync").}

\item{instr}{Instrument name to be processed}

\item{beat}{Variable name where the beats are}

\item{virtual}{Variable name where the virtual beats are}

\item{pcols}{Number of columns for multiple plots (optional)}

\item{griddeviations}{Add deviation from the virtual beats  (optional)}

\item{boxplot}{Do the graphics by boxplot  (optional)}

\item{colour}{colour for the boxplot  (optional)}

\item{colourpalette}{colors for dots in the timeline, use 'grey' for bw
(optional)}

\item{pointsize}{size of the dots (defaults 1)}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels
these by instruments.
}
