% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_pair.R
\name{plot_by_pair}
\alias{plot_by_pair}
\title{Plot synchronies by instrument pairs}
\usage{
plot_by_pair(df = NULL, bybeat = FALSE, reference = 0, colourpalette = "Set2")
}
\arguments{
\item{df}{data frame to be processed}

\item{bybeat}{Optional (FALSE by default) which adds the beats to the plot}

\item{reference}{Optional (0 by default) which adds the reference line to the plot}

\item{colourpalette}{Optional colour palette, Default 'PuOr' (Purple to Orange)}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instrument pairs.
}
\seealso{
\code{\link{plot_by_variable}}
}
