% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_var_time.R
\name{plot_by_var_time}
\alias{plot_by_var_time}
\title{Plots a variables across time and possible by any other structure}
\usage{
plot_by_var_time(
  df,
  var1 = "Time",
  var2 = "Synchrony",
  var3 = NULL,
  colour = "orange",
  smoothline_colour = "navyblue",
  xlabel = NULL,
  ylabel = NULL,
  zlabel = NULL
)
}
\arguments{
\item{df}{data frame to be processed}

\item{var1}{Variable 1 (Time)}

\item{var2}{Variable 2}

\item{var3}{Variable 3 (optional)}

\item{colour}{Colour option}

\item{smoothline_colour}{Colour option}

\item{xlabel}{label for X axis}

\item{ylabel}{label for Y axis}

\item{zlabel}{label for Z axis}
}
\value{
Graphic output
}
\description{
\code{plot_by_var_time} visualizes the
calculated asynchronies of instruments across time and
another variable.
}
\seealso{
\code{\link{sync_sample_paired}} for synchronies between
instruments, \code{\link{plot_by_beat}} for plotting.
}
