% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_variable.R
\name{plot_by_variable}
\alias{plot_by_variable}
\title{Plots synchronies by any other structure}
\usage{
plot_by_variable(df, meta = "empty", xlab = "Tempo")
}
\arguments{
\item{df}{data frame to be processed (required)}

\item{meta}{Variable information (Pair of instrument names, for instance)}

\item{xlab}{label for X axis}
}
\value{
Graphic output
}
\description{
\code{plot_by_variable} visualizes the 
calculated asynchronies of instruments across
another variable.
}
\seealso{
\code{\link{sync_sample_paired}} for synchronies between 
instruments, \code{\link{plot_by_beat}} for plotting.
}
