% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Plot onsets across time for instruments.}
\usage{
plot_timeline(data = NULL, instr = NULL, ampl = NULL, grouping = NULL)
}
\arguments{
\item{data}{data frame to be processed}

\item{instr}{Instrument name to be processed}

\item{ampl}{Amplitude of the onsets (optional)}

\item{grouping}{Grouping factor (optional)}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
\seealso{
\code{\link{plot_by_variable}}
}
