% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_periodicity.R
\name{summarise_periodicity}
\alias{summarise_periodicity}
\title{Summarise onset periodicities by calculating the moments (mean, etc.)}
\usage{
summarise_periodicity(data = NULL, method = "max")
}
\arguments{
\item{data}{data frame to be processed}

\item{method}{method of choosing the peaks (max default, or first)}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
\seealso{
\code{\link{periodicity}}
}
