% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_sync.R
\name{summarise_sync}
\alias{summarise_sync}
\title{Summarise synchronies}
\usage{
summarise_sync(df = NULL)
}
\arguments{
\item{df}{data frame to be processed (required)}
}
\value{
Integer output
}
\description{
This function summarises the calculated asynchronies.
}
