% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_sync_by_pair.R
\name{summarise_sync_by_pair}
\alias{summarise_sync_by_pair}
\title{Calculate summary statistics for asynchronies by instrument pairs}
\usage{
summarise_sync_by_pair(df, bybeat = FALSE, adjust = "bonferroni")
}
\arguments{
\item{df}{data frame to be processed}

\item{bybeat}{logical value whether beat information is used}

\item{adjust}{Adjust p values with Bonferroni correction (default)}
}
\value{
Statistics output (Mean, Standard deviation, T-stat, P value)
}
\description{
This function calculates simple t-test of the asynchronies of instruments (whether it differs from zero)
}
\seealso{
\code{\link{sync_sample_paired}}, \code{\link{sync_execute_pairs}}
}
