% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_measures.R
\name{summary_measures}
\alias{summary_measures}
\title{Plot synchronies by instrument pairs}
\usage{
summary_measures(df = NULL, relative = FALSE)
}
\arguments{
\item{df}{data frame to be processed}

\item{relative}{Whether calculation is relative measure}
}
\value{
Integer output
}
\description{
This function plots the calculated asynchronies of instrument pairs.
}
