% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_joint_onsets.R
\name{sync_joint_onsets}
\alias{sync_joint_onsets}
\title{Determine the number of joint onsets for two instruments}
\usage{
sync_joint_onsets(df = NULL, instr1 = NULL, instr2 = NULL)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr1}{Instrument 1 name to be processed}

\item{instr2}{Instrument 2 name to be processed}
}
\value{
List containing asynchronies and beat structures
}
\description{
Calculate the number of joint onsets for two instruments
}
\seealso{
\code{\link{sync_execute_pairs}}
}
