% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_sample_paired_relative.R
\name{sync_sample_paired_relative}
\alias{sync_sample_paired_relative}
\title{Compare instrument onset times to mean onset times from several instruments}
\usage{
sync_sample_paired_relative(
  df = NULL,
  instr = NULL,
  instr_ref = NULL,
  beat = NULL
)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{Instrument name to be processed}

\item{instr_ref}{Instrument reference names to be processed}

\item{beat}{Beat structure (subdivisions) to be included}
}
\value{
difference in ms
}
\description{
Compare instrument onset times to mean onset times from several instruments
}
\seealso{
\code{\link{sync_execute_pairs}}
}
