% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise_onsets.R
\name{synthesise_onsets}
\alias{synthesise_onsets}
\title{Synthesise onsets with noises and clicks}
\usage{
synthesise_onsets(
  data = NULL,
  instruments = NULL,
  sr = 22050,
  dur = 0.03,
  type = "noise",
  freq = 2500
)
}
\arguments{
\item{data}{data frame to be processed (required)}

\item{instruments}{Instruments names to be included (required)}

\item{sr}{sampling rate (default 22050)}

\item{dur}{duration of sounds (default 30 ms)}

\item{type}{type of sounds (default noise, try click or synth from seewave)}

\item{freq}{freq of sounds (if synth is given as input)}
}
\value{
wave file
}
\description{
Take the specified instruments and create wave file with sounds where the onsets are
}
