% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussify_onsets.R
\name{gaussify_onsets}
\alias{gaussify_onsets}
\title{Create a continuous vector of onsets with a sampling in Hz}
\usage{
gaussify_onsets(data = NULL, sr = 250, wlen = 0.2, plot = FALSE, time = TRUE)
}
\arguments{
\item{data}{Data frame to be processed}

\item{sr}{Sampling rate (default 50 Hz)}

\item{wlen}{Window length of percentage of the sr}

\item{plot}{If a plot is needed}

\item{time}{Time samples to be included in the output}
}
\value{
Output contain vector of onset times
}
\description{
This is a conversion function that turns onsets into sampled time-series.
}
