% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_by_beat_plot.R
\name{onsetsynch_by_beat_plot}
\alias{onsetsynch_by_beat_plot}
\title{plot synchronies by beat structures}
\usage{
onsetsynch_by_beat_plot(df, instr, beat, virtual, pcols)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{Instrument name to be processed}

\item{beat}{Variable name where the beats are}

\item{virtual}{Variable name where the virtual beats are}

\item{pcols}{Number of columns for multiple plots}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
