% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onsetsynch_by_X_plot.R
\name{onsetsynch_by_X_plot}
\alias{onsetsynch_by_X_plot}
\title{Plots synchronies by any other structure}
\usage{
onsetsynch_by_X_plot(df, meta = "empty", xlab = "Tempo")
}
\arguments{
\item{df}{data frame to be processed}

\item{meta}{Variable information (Pair of instrument names, for instance)}

\item{xlab}{label for X axis}
}
\value{
Graphic output
}
\description{
\code{onsetsynch_by_X_plot} visualises the 
calculated asynchronies of instruments across
another variable.
}
\seealso{
\code{\link{onsetsynch_sample_paired}} for synchronies between 
instruments, \code{\link{onsetsynch_by_beat_plot}} for plotting.
}
