% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity.R
\name{periodicity}
\alias{periodicity}
\title{Estimate the periodicity of the onsets by a specific method}
\usage{
periodicity(
  df = NULL,
  instr = NULL,
  method = "acf",
  sampling_rate = 250,
  freq_range = c(0, 2),
  resolution = 0.01,
  ignore_period_under = 0.2,
  ignore_period_above = 5,
  colour = "navyblue",
  title = NULL
)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{Instrument name to be processed}

\item{method}{Periodicity analysis method (diff, acf, fft, wavelet)}

\item{sampling_rate}{Sampling rate (Hz)}

\item{freq_range}{Frequency range to be included in the periodicity analysis (in seconds)}

\item{resolution}{Resolution for some of the analyses (in seconds)}

\item{ignore_period_under}{Ignore period lower than this (in seconds)}

\item{ignore_period_above}{Ignore period above than this (in seconds)}

\item{colour}{Line colour for plotting}

\item{title}{Title for plotting}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
\seealso{
\code{\link{periodicity_moments}}

\code{\link{periodicity_moments}} for summary function.
}
