% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_beat.R
\name{plot_by_beat}
\alias{plot_by_beat}
\title{Plot synchronies by beat structures}
\usage{
plot_by_beat(
  df = NULL,
  instr = NULL,
  beat = NULL,
  virtual = NULL,
  pcols = 1,
  griddeviations = FALSE,
  boxplot = FALSE,
  colour = "lightblue",
  colourpalette = "Set1",
  pointsize = 1
)
}
\arguments{
\item{df}{data frame to be processed}

\item{instr}{Instrument name to be processed}

\item{beat}{Variable name where the beats are}

\item{virtual}{Variable name where the virtual beats are}

\item{pcols}{Number of columns for multiple plots}

\item{griddeviations}{Add deviation from the virtual beats in}

\item{boxplot}{Do the graphics by boxplot}

\item{colour}{colour for the boxplot}

\item{colourpalette}{colors for dots in the timeline, use 'grey' for bw}

\item{pointsize}{size of the dots (defaults 1)}
}
\value{
Graphic output
}
\description{
This function plots the calculated asynchronies of instruments and labels these by instruments.
}
