% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_onsets.R
\name{summarise_onsets}
\alias{summarise_onsets}
\title{Summarise onsets for instruments specified}
\usage{
summarise_onsets(
  df = NULL,
  instr = NULL,
  filter_lower = 0,
  filter_upper = 2,
  binw = 0.25,
  plot = FALSE
)
}
\arguments{
\item{df}{data frame to be processed (required)}

\item{instr}{Instrument names to be processed  (required)}

\item{filter_lower}{Lower limit for onset differences to be summarised (default 0)}

\item{filter_upper}{Upper limit for onset differences to be summarised (default 2)}

\item{binw}{bin width for histogram (default 0.25)}

\item{plot}{Option for plotting (defaul FALSE)}
}
\value{
Output contain summarises of the onset differences
}
\description{
\code{summarise_onsets} is a function that
summarises the time difference between successive onsets
across instruments.
}
