% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_execute_pairs.R
\name{sync_execute_pairs}
\alias{sync_execute_pairs}
\title{Calculate asynchronies across all pairs of instruments}
\usage{
sync_execute_pairs(
  df = NULL,
  instruments = NULL,
  n_samp = 100,
  bnum = 1,
  beats = NULL
)
}
\arguments{
\item{df}{data frame to be processed}

\item{instruments}{Instrument names to be processed}

\item{n_samp}{Number of samples to be drawn from the pool of joint onsets}

\item{bnum}{How many bootstraps are drawn}

\item{beats}{Beat structure to be included}
}
\value{
Output contain asynchronies and beat levels
}
\description{
\code{sync_execute_pairs} is a meta-function that
executes calculations of asynchronies across all pairs of
instruments. One can specify the same options as in
\code{sync_sample_paired}.
}
\seealso{
\code{\link{sync_sample_paired}}
}
