% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_execute_pairs.R
\name{sync_execute_pairs}
\alias{sync_execute_pairs}
\title{Calculate asynchronies across all pairs of instruments}
\usage{
sync_execute_pairs(
  df = NULL,
  instruments = NULL,
  n = 0,
  bootn = NULL,
  beat = NULL
)
}
\arguments{
\item{df}{data frame to be processed, where the minimal requirements is
two or more instruments (instruments)}

\item{instruments}{Two of more instruments (required)}

\item{n}{Number of samples to be drawn from the pool of joint onsets (default 0)}

\item{bootn}{How many bootstraps are drawn (default=NULL)}

\item{beat}{Beat structure (subdivisions) to be included (default=NULL)}
}
\value{
Output containing asynchronies and beat levels
}
\description{
\code{sync_execute_pairs} is a meta-function that
executes calculations of asynchronies across all pairs of
instruments. One can specify the same options as in
\code{sync_sample_paired}.
}
\seealso{
\code{\link{sync_sample_paired}}
}
