% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_sample_paired.R
\name{sync_sample_paired}
\alias{sync_sample_paired}
\title{Take equal number of samples from two instruments}
\usage{
sync_sample_paired(
  df = NULL,
  INSTR1 = NULL,
  INSTR2 = NULL,
  N = 100,
  BNum = 1,
  beat = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{data frame to be processed}

\item{INSTR1}{Instrument 1 name to be processed}

\item{INSTR2}{Instrument 2 name to be processed}

\item{N}{Number of samples to be drawn from the pool of joint onsets. If 0, do not sample!}

\item{BNum}{How many bootstraps are drawn}

\item{beat}{Beat structure to be included}

\item{verbose}{Display no. of shared onsets (default FALSE)}
}
\value{
List containing asynchronies and beat structures
}
\description{
taken N samples of two instruments (where they both have onsets)
}
\seealso{
\code{\link{sync_execute_pairs}}
}
